static library impl
