+++ .idea/gradle.xml
@@ -9,6 +9,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/v2pxcalendar" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
