+++ v2pxcalendar/src/main/java/com/rosia/v2pxcalendar/ui/adapter/CalendarItemDecoration.kt
@@ -1,10 +1,10 @@
-package com.rosia.calendartest
+package com.rosia.v2pxcalendar.ui.adapter
 
 import android.content.Context
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import com.rosia.calendartest.adapter.CalendarDateAdapter
+import com.rosia.v2pxcalendar.ui.adapter.CalendarDateAdapter
 
 class CalendarItemDecoration(var context: Context): RecyclerView.ItemDecoration(){
 	override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
