+++ v2pxcalendar/src/main/java/com/rosia/v2pxcalendar/ui/models/CustomDay.kt
@@ -1,4 +1,4 @@
-package com.rosia.calendartest.models
+package com.rosia.v2pxcalendar.ui.models
 
 data class CustomDay(
 	val day: String,
