+++ v2pxcalendar/src/main/res/drawable/bg_rec_grey.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle">
+
+	<size
+		android:width="30dp"
+		android:height="30dp" />
+
+	<solid android:color="@color/grey_FFF3F4F6" />
+
+</shape>
