+++ v2pxcalendar/src/main/res/drawable/bg_rounded_end_blue.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle">
+
+	<size
+		android:width="30dp"
+		android:height="30dp" />
+
+	<corners
+		android:bottomRightRadius="20dp"
+		android:topRightRadius="20dp" />
+
+	<solid android:color="@color/blue_FF1294F7" />
+
+</shape>
