+++ v2pxcalendar/src/main/res/layout/item_date.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:id="@+id/item_container"
+	android:layout_width="match_parent"
+	android:layout_height="30dp"
+	android:layout_marginTop="2dp"
+	android:layout_marginBottom="2dp">
+
+
+	<TextView
+		android:id="@+id/text_date"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:gravity="center"
+		android:textAllCaps="false"
+		android:textColor="@color/black_2e384d"
+		android:textSize="12sp"
+		app:layout_constraintBottom_toBottomOf="parent"
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintHorizontal_bias="0.529"
+		app:layout_constraintStart_toStartOf="parent"
+		app:layout_constraintTop_toTopOf="parent"
+		tools:text="13" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
+
+
