+++ v2pxcalendar/src/main/res/layout/view_calendar.xml
@@ -1,0 +1,78 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	android:background="@color/white_FFFFFF">
+
+	<ImageView
+		android:id="@+id/btn_previous"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_marginStart="24dp"
+		android:layout_marginTop="16dp"
+		app:layout_constraintStart_toStartOf="parent"
+		app:layout_constraintTop_toTopOf="parent"
+		app:srcCompat="@drawable/ic_arrow_left" />
+
+	<ImageView
+		android:id="@+id/btn_next"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_marginTop="16dp"
+		android:layout_marginEnd="24dp"
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintTop_toTopOf="parent"
+		app:srcCompat="@drawable/ic_arrow_right" />
+
+	<TextView
+		android:id="@+id/textMonthYear"
+		android:layout_width="0dp"
+		android:layout_height="wrap_content"
+		android:layout_marginStart="8dp"
+		android:layout_marginEnd="8dp"
+		android:textAlignment="center"
+		android:textColor="@color/black_2e384d"
+		app:layout_constraintBottom_toBottomOf="@+id/btn_previous"
+		app:layout_constraintEnd_toStartOf="@+id/btn_next"
+		app:layout_constraintStart_toEndOf="@+id/btn_previous"
+		app:layout_constraintTop_toTopOf="@+id/btn_previous"
+		tools:text="July 2019" />
+
+	<androidx.recyclerview.widget.RecyclerView
+		android:id="@+id/recyclerview_day"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:layout_marginStart="16dp"
+		android:layout_marginTop="16dp"
+		android:layout_marginEnd="16dp"
+		android:overScrollMode="never"
+		android:scrollbars="none"
+		app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintStart_toStartOf="parent"
+		app:layout_constraintTop_toBottomOf="@+id/textMonthYear"
+		app:spanCount="7"
+		tools:itemCount="7"
+		tools:listitem="@layout/item_day" />
+
+	<androidx.recyclerview.widget.RecyclerView
+		android:id="@+id/recycler_date"
+		android:layout_width="0dp"
+		android:layout_height="wrap_content"
+		android:layout_marginStart="16dp"
+		android:layout_marginTop="8dp"
+		android:layout_marginEnd="16dp"
+		android:layout_marginBottom="8dp"
+		android:overScrollMode="never"
+		app:layout_constraintBottom_toBottomOf="parent"
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintStart_toStartOf="parent"
+		app:layout_constraintTop_toBottomOf="@+id/recyclerview_day"
+		app:layout_constraintVertical_bias="0.0"
+		tools:itemCount="30"
+		tools:listitem="@layout/item_date" />
+</androidx.constraintlayout.widget.ConstraintLayout>
